% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completePairSubset.R
\name{completePairSubset}
\alias{completePairSubset}
\title{Function that runs detects complete pairs}
\usage{
completePairSubset(seuratObject, uniqueSampleID = "sample", pairingColumn)
}
\arguments{
\item{seuratObject}{Seurat v5 Object}

\item{uniqueSampleID}{Column in seurat metadata object of the unique sample ID}

\item{pairingColumn}{Column in seurat metadata object that samples are paired on}
}
\value{
seuratObject of data with complete pairs
}
\description{
Function that runs detects complete pairs
}
