% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirmPairs.R
\name{confirmPairs}
\alias{confirmPairs}
\title{Function that runs detects complete pairs}
\usage{
confirmPairs(seuratObject, uniqueSampleID = "sample", pairingColumn)
}
\arguments{
\item{seuratObject}{Seurat v5 Object}

\item{uniqueSampleID}{Column in seurat metadata object of the unique sample ID}

\item{pairingColumn}{Column in seurat metadata object that samples are paired on}
}
\value{
boolean TRUE if two samples per pair present, FALSE if not
}
\description{
Function that runs detects complete pairs
}
