% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runscMI.R
\name{formatMI}
\alias{formatMI}
\title{Function that converts a list of MetaIntegrator datasetObjects into a single MetaIntegrator metaObject}
\usage{
formatMI(listOfMetaIntegratorObjs)
}
\arguments{
\item{listOfMetaIntegratorObjs}{list of MetaIntegrator datasetObjects}
}
\value{
MetaIntegrator metaObject
}
\description{
Function that converts a list of MetaIntegrator datasetObjects into a single MetaIntegrator metaObject
}
