% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runscMI.R
\name{generateFullMA}
\alias{generateFullMA}
\title{Function that runs the full scMI pipeline and outputs complete metaObject}
\usage{
generateFullMA(
  seuratObject,
  uniqueSampleID = "sample",
  pairingColumn,
  comparisonGroupColumn,
  groupOne = "Condition",
  groupTwo = "Control",
  thresholdGenes = NULL,
  completePairsOnly = FALSE
)
}
\arguments{
\item{seuratObject}{Seurat v5 Object}

\item{uniqueSampleID}{Column in seurat metadata object of the unique sample ID}

\item{pairingColumn}{Column in seurat metadata object that samples are paired on}

\item{comparisonGroupColumn}{Column in seurat metadata object to use for DE comparisons}

\item{groupOne}{Identity in comparison group to define DEGs}

\item{groupTwo}{A second identity for comparison, ie control group}

\item{thresholdGenes}{Can include a vector of genes to include for analysis. If left empty, all genes will be considered}

\item{completePairsOnly}{BOOL where TRUE subsets seuratObject to samples from only the complete pairs}
}
\value{
metaObject of the processed analyses
}
\description{
Function that runs the full scMI pipeline and outputs complete metaObject
}
