% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runscMI.R
\name{seuToMI}
\alias{seuToMI}
\title{Function that convert seuratObject to MetaIntegrator datasetObject}
\usage{
seuToMI(
  seuratObject,
  thresholdGenes = NULL,
  groupOne,
  groupTwo,
  comparisonGroupColumn = "condition"
)
}
\arguments{
\item{seuratObject}{Seurat v5 Object}

\item{thresholdGenes}{Can include a vector of genes to include for analysis. If left empty, all genes will be considered}

\item{groupOne}{Identity in comparison group to define DEGs}

\item{groupTwo}{A second identity for comparison, ie control group}

\item{comparisonGroupColumn}{Column in seurat metadata object to use for DE comparisons}
}
\value{
seuratObject of data with complete pairs
}
\description{
Function that convert seuratObject to MetaIntegrator datasetObject
}
