% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdGenesAcrossPairs.R
\name{thresholdGenesAcrossPairs}
\alias{thresholdGenesAcrossPairs}
\title{Function that runs thresholding of genes by sample pairs}
\usage{
thresholdGenesAcrossPairs(
  seuratObject,
  uniqueSampleID = "sample",
  pairingColumn,
  percentThreshold = 0.01,
  completePairsOnly = FALSE,
  assay = "RNA"
)
}
\arguments{
\item{seuratObject}{Bayesian Meta-Analysis object with basic bayesian meta-analysis run}

\item{uniqueSampleID}{Column in seurat metadata object of the unique sample ID}

\item{pairingColumn}{Column in seurat metadata object that samples are paired on}

\item{percentThreshold}{Percentage of cells that should express the gene across each pair to be included}

\item{completePairsOnly}{BOOL where TRUE subsets seuratObject to samples from only the complete pairs}

\item{assay}{assay to use}
}
\value{
Vector of genes that pass thresholds
}
\description{
Function that runs thresholding of genes by sample pairs
}
